Add event listeners to buttons to console.log user choice
I want the buttons to select the user's choice of rock, paper, or scissors. I think this is an important step towards making the buttons functional -- they need to grab the player's choice, which can then be used when the round runs. I think I need to add the event listener for the playRound function next, per the instructions, and have it pull in whatever choice is being printed to the console for the user's choice. So far, it's outputting null.