Edit playRound function to work with event listener
I had to remove the getHumanChoice function and let the button clicks float in the global scope. I added a parameter to playRound to allow for a string representing the button's input ("ROCK", "PAPER", or "SCISSORS") to be passed as an argument and used to compare with the computer's choice. Now, when a button is clicked, it passes a string representing that button into playRound and calls playRound/logs it to the console.