Reverted changes to playRound and computerChose
Moving the computerSelection variable out of the function made it no longer run whenever the function was called, so it was only ever picking one option for the computer. I need to figure out how to have the variable available for the event listener I want to add to the buttons. Maybe I add this event listener to the function so it can use the computerSelection variable.